Restructured the navbar component
- split it up into more files for ease of coding/debugging
