+++ client/src/components/NavBar/index.jsx
@@ -1,0 +1,3 @@
+import NavBar from "./NavBar";
+
+export default NavBar;
